Fix auth tests for new endpoints
POST is not allowed on collection endpoints.
It was allowed before for the case when the class endpoint is
of a class which does not form a collection. But in the new
endpoint system, there is no need of this differentiation as
a class which does not form a collection and which does form a collection,
both have same behavior with respect to endpoints.

So changed POST to PUT.