Better discover collections (#483)
* Remove hardcoding of "Collection" keyword

* Remove dependency of collection word in crud.py

All the resources uris for collections will now be made just from
path variable information. Their will not be dependency
of 'type_' variable on resource uri. As the 'type_' value
might be something and the actual 'path' for that collection
can be different. It need not be always 'type_Collection'

* Fix @type in pagination

* Fix the condition for non-collection classes